Make Interval size inclusive
By default I'm making Interval a *closed* interval.  The endpoints are
included in the set.  I've decided that this is the most intuitive
interpretation and will be the default for Ambit.  In the future I
would like to explore the possibility of a notation for open intervals
as well as left/right exclusions...